CORE-345: Fill out ‘ImplC’; implement BRSystem.swift using ‘ImplC’
