+++ Swift/Core/module.modulemap
@@ -10,7 +10,7 @@ module BRCryptoC [system] [extern_c] {
     header "../../crypto/BRCryptoFeeBasis.h"
     header "../../crypto/BRCryptoTransfer.h"
     header "../../crypto/BRCryptoWallet.h"
-//    header "../../crypto/BRCryptoWalletManager.h"
+    header "../../crypto/BRCryptoWalletManager.h"
 //    header "../../crypto/BRCryptoSystem.h"
 
     header "../../crypto/BRCryptoPrivate.h"
